refactor(app): fold image/audio scan-directory handlers into useAssetManagement
Moves handleAdd/Refresh/RemoveScanDirectory and handleCopy*ToProjectBulk
for both images and audio out of App.tsx into useAssetManagement, using
the same params-object pattern as useFileSystemManager/useProjectIO for
cross-cutting deps (perfRecorders, addToast, setHasUnsavedSettings, etc).
App.tsx: 4216 -> 4029 lines.
