+++ .beads/interactions.jsonl
@@ -27,3 +27,4 @@
 {"id":"int-d208ad6f","kind":"field_change","created_at":"2026-06-23T03:01:32.2470042Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-gn2","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Extracted loadProject/handleSaveProjectSettings/handleSaveAll/handleReloadFromDisk/handleRefreshProject into src/hooks/useProjectIO.ts. App.tsx 5314->4598 lines (-716). tsc/lint/686 tests/build all green, no regressions (verified via before/after diff of pre-existing tsc errors)."}}
 {"id":"int-665fe78e","kind":"field_change","created_at":"2026-06-23T03:11:23.7711819Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-13l","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Moved handleCreateNode/handleRenameNode/handleDeleteNode/handleMoveNode/handleCut/handleCopy/handlePaste into useFileSystemManager.ts alongside the existing tree-helper primitives. App.tsx 4598->4480 lines (-118). tsc/lint/686 tests/build all green, no regressions."}}
 {"id":"int-7fc1e6ff","kind":"field_change","created_at":"2026-06-23T11:20:36.6422212Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-4oq","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Extracted getTabLabel/renderTabContent/renderTabBar into src/hooks/useTabContentRenderer.tsx via props-bag pattern matching useProjectIO/useFileSystemManager. App.tsx reduced 4480->4216 lines. tsc/lint/tests all clean, no new errors vs baseline, 686/686 tests pass."}}
+{"id":"int-d8030abb","kind":"field_change","created_at":"2026-06-23T11:56:03.2696758Z","actor":"G.Hirpara","issue_id":"bmf-vangard-renpy-ide-2uu","extra":{"field":"status","new_value":"closed","old_value":"in_progress","reason":"Folded handleAdd/Refresh/Remove ScanDirectory and CopyToProjectBulk for images+audio into useAssetManagement (params-object pattern: projectRootPath, perfRecorders, setIsScanningAssets, setHasUnsavedSettings, setFileSystemTree, addToast). App.tsx 4216->4029 lines (-187). Tests 686/686 pass, lint clean, build succeeds, tsc baseline error count unchanged (50)."}}
